Add new VAT API client
